combine bytes and files test files
